Convert to use X11
This is a pretty big change in how this program works, now that I'm
using X11 instead of raw framebuffer/DRI. Now, this program modifies an
xorg config file that contains the calibration matrix for libinput.

This is nice because it also modifies the property live, so we won't
need to restart the X server in order to apply the new calibration.

+++ calibrationwindow.h
@@ -54,9 +54,9 @@ private:
     bool _touchIsPressed;
     QQueue<QPoint> _points;
 
-    int _minXCal;
-    int _maxXCal;
-    int _minYCal;
-    int _maxYCal;
+    float _minXCal;
+    float _maxXCal;
+    float _minYCal;
+    float _maxYCal;
 };
 #endif // CALIBRATIONWINDOW_H
